move system piping tests into `mod.rs`
